🩹 [Patch]: Update linter and git configurations and license year (#24)
## Description

This pull request includes several changes to the configuration files
for various linters and an update to the license year. The most
important changes are grouped into updates to linter configurations and
a minor license update.

Linter configuration updates:

*
[`.github/linters/.jscpd.json`](diffhunk://#diff-557094e283c00b23265c1c75872f41c6b1a524a00f0d99dd68ebd22cb63bfdd6R1-R10):
Added a new configuration file to set up the JavaScript Copy/Paste
Detector (jscpd) with specific settings including a threshold of 0 and
ignoring test files.
*
[`.github/linters/.powershell-psscriptanalyzer.psd1`](diffhunk://#diff-aae69c9d6774628ed181eacf53aee0f38eb6c2f53492cf3a5b7f7bdb6ef43b6aL1-L16):
Replaced the existing configuration with new rules for PowerShell Script
Analyzer, including rules for alignment, line length, semicolons,
braces, comments, indentation, and whitespace.
*
[`.github/linters/.markdown-lint.yml`](diffhunk://#diff-2fd3619d2b7372123257bc21645edb90c24293e66babbc6053033a89e21bab34L9-R17):
Made minor adjustments to the markdown lint configuration, such as
changing the case of a comment and adjusting punctuation settings.
*
[`.github/workflows/Linter.yml`](diffhunk://#diff-482e65806ed9e4a7320f14964764086b91fed4a28d12e4efde1776472e147e79R30):
Added an environment variable to disable JSON Prettier validation in the
GitHub Actions workflow for the linter.

Minor update:

*
[`LICENSE`](diffhunk://#diff-c693279643b8cd5d248172d9c22cb7cf4ed163a3c98c8a3f69c2717edd3eacb7L3-R3):
Updated the copyright year from 2024 to 2025.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas