🩹 [Patch]: Update FontsData patcher to use separate app and run on a schedule (#25)
## Description

This pull request includes updates to the `Update-FontsData` workflow
and script to automate the process of updating font data and opening a
pull request if changes are detected. The most important changes include
switching to a scheduled workflow, adding authentication for the GitHub
App, and enhancing the script to handle branch creation, commit, and PR
creation.

Improvements to workflow automation:

*
[`.github/workflows/Update-FontsData.yml`](diffhunk://#diff-5c6d6ac6744e613c22fe7a4b5bf6b9fbd4b3bf5548239191e4400a4b17440826L4-R10):
Changed the trigger to `workflow_dispatch` and added a schedule to run
daily at midnight. Updated permissions to allow writing pull requests.

Authentication and script enhancements:

*
[`.github/workflows/Update-FontsData.yml`](diffhunk://#diff-5c6d6ac6744e613c22fe7a4b5bf6b9fbd4b3bf5548239191e4400a4b17440826L22-R25):
Updated the `Update-FontsData` job to include `ClientID` and
`PrivateKey` for authentication.
*
[`scripts/Update-FontsData.ps1`](diffhunk://#diff-69d235ffacf154789a70d5e00a557ee83393c256cd9431a8b7be57cc41c3ad89L1-R14):
Added steps to connect to the GitHub App, create a new branch, and
retrieve the latest font data.
*
[`scripts/Update-FontsData.ps1`](diffhunk://#diff-69d235ffacf154789a70d5e00a557ee83393c256cd9431a8b7be57cc41c3ad89R25-R52):
Enhanced the script to write results to `FontsData.json`, check for
changes, commit, and push changes, and create a pull request if changes
are detected.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas