🪲 [Fix]: Remove preserve-credential on checkout (#27)
## Description

This pull request includes changes to the `Update-FontsData` workflow
and script to improve security and functionality. The most important
changes include updating the GitHub Actions workflow to avoid persisting
credentials and modifying the PowerShell script to use a default
connection for the GitHub App.

Improvements to security and functionality:

*
[`.github/workflows/Update-FontsData.yml`](diffhunk://#diff-5c6d6ac6744e613c22fe7a4b5bf6b9fbd4b3bf5548239191e4400a4b17440826R17-R18):
Updated the `Checkout repository` step to avoid persisting credentials
by setting `persist-credentials` to `false`.
*
[`scripts/Update-FontsData.ps1`](diffhunk://#diff-69d235ffacf154789a70d5e00a557ee83393c256cd9431a8b7be57cc41c3ad89L1-R1):
Modified the `Connect-GitHubApp` command to use the `-Default` parameter
for a default connection.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas